+++ Source/Libraries/Core/Business/Vodovoz.Reports/Reports/Logistic/DeliveriesLate.rdl
@@ -11,7 +11,7 @@
   <BottomMargin>5mm</BottomMargin>
   <DataSets>
     <DataSet Name="Data">
-      <Query>
+      <Query >
         <DataSourceName>DS1</DataSourceName>
         <CommandText>SELECT
     route_lists.id,
@@ -46,11 +46,6 @@ FROM
         INNER JOIN employees driver ON
                 driver.id = route_lists.driver_id
                 AND driver.visiting_master = 0
-                AND
-                    ((@exclude_truck_drivers_office_employees = 1
-                        AND driver.category != 'office'
-                        AND driver.driver_of_car_type_of_use != 'Truck')
-                    OR @exclude_truck_drivers_office_employees = 0)
         LEFT JOIN route_list_addresses ON route_list_addresses.route_list_id = route_lists.id
         LEFT JOIN orders ON orders.id = route_list_addresses.order_id
         LEFT JOIN counterparty ON counterparty.id = orders.client_id
@@ -77,9 +72,6 @@ HAVING delay &gt; 0;</CommandText>
           <QueryParameter Name="geographic_group_id">
             <Value>={?geographic_group_id}</Value>
           </QueryParameter>
-          <QueryParameter Name="exclude_truck_drivers_office_employees">
-            <Value>={?exclude_truck_drivers_office_employees}</Value>
-          </QueryParameter>
           <QueryParameter Name="order_select_mode">
             <Value>={?order_select_mode}</Value>
           </QueryParameter>
@@ -88,7 +80,7 @@ HAVING delay &gt; 0;</CommandText>
           </QueryParameter>
         </QueryParameters>
       </Query>
-      <Fields>
+      <Fields >
         <Field Name="id">
           <DataField>id</DataField>
           <rd:TypeName>System.UInt32</rd:TypeName>
@@ -174,7 +166,7 @@ WHERE o.order_status != 'NotDelivered' and o.order_status != 'Canceled'
       </Fields>
     </DataSet>
     <DataSet Name="LatenessSummary">
-      <Query >
+      <Query>
         <DataSourceName>DS1</DataSourceName>
         <CommandText>SELECT
 	Sum(Delays.delay &lt; '00:05:00') AS LessThan5Minutes,
@@ -246,7 +238,7 @@ WHERE o.order_status != 'NotDelivered' and o.order_status != 'Canceled'
           </QueryParameter>
         </QueryParameters>
       </Query>
-      <Fields >
+      <Fields>
         <Field Name="LessThan5Minutes">
           <DataField>LessThan5Minutes</DataField>
           <TypeName>System.String</TypeName>
